requirements.txt
Django==5.0.6
granian==1.4.1
pip~=23.2.1
asgiref~=3.8.1
sqlparse~=0.5.0
uvloop~=0.18.0
click~=8.1.7
ujson~=5.10.0
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
psycopg==3.1.13
psycopg-binary==3.1.13
bjoern==3.2.2
uwsgi==2.0.26
